+++ R/h_survival_duration_subgroups.R
@@ -220,7 +220,7 @@ h_coxph_df <- function(tte, is_event, arm, strata_data = NULL, control = control
       .in_ref_col = FALSE,
       .var = "tte",
       is_event = "is_event",
-      strat = strata_vars,
+      strata = strata_vars,
       control = control
     )
 
@@ -277,7 +277,7 @@ h_coxph_df <- function(tte, is_event, arm, strata_data = NULL, control = control
 #' @describeIn h_survival_duration_subgroups summarizes estimates of the treatment hazard ratio
 #'   across subgroups in a data frame. `variables` corresponds to the names of variables found in
 #'   `data`, passed as a named list and requires elements `tte`, `is_event`, `arm` and
-#'   optionally `subgroups` and `strat`. `groups_lists` optionally specifies
+#'   optionally `subgroups` and `strata`. `groups_lists` optionally specifies
 #'   groupings for `subgroups` variables.
 #'
 #' @return
@@ -321,7 +321,7 @@ h_coxph_df <- function(tte, is_event, arm, strata_data = NULL, control = control
 #'     is_event = "is_event",
 #'     arm = "ARM",
 #'     subgroups = c("SEX", "BMRKR2"),
-#'     strat = c("STRATA1", "STRATA2")
+#'     strata = c("STRATA1", "STRATA2")
 #'   ),
 #'   data = adtte_f
 #' )
@@ -332,11 +332,20 @@ h_coxph_subgroups_df <- function(variables,
                                  groups_lists = list(),
                                  control = control_coxph(),
                                  label_all = "All Patients") {
+  if ("strat" %in% names(variables)) {
+    warning(
+      "Warning: the `strat` element name of the `variables` list argument to `h_coxph_subgroups_df() ",
+      "was deprecated in tern 0.9.3.\n  ",
+      "Please use the name `strata` instead of `strat` in the `variables` argument."
+    )
+    variables[["strata"]] <- variables[["strat"]]
+  }
+
   checkmate::assert_character(variables$tte)
   checkmate::assert_character(variables$is_event)
   checkmate::assert_character(variables$arm)
   checkmate::assert_character(variables$subgroups, null.ok = TRUE)
-  checkmate::assert_character(variables$strat, null.ok = TRUE)
+  checkmate::assert_character(variables$strata, null.ok = TRUE)
   assert_df_with_factors(data, list(val = variables$arm), min.levels = 2, max.levels = 2)
   assert_df_with_variables(data, variables)
   checkmate::assert_string(label_all)
@@ -346,7 +355,7 @@ h_coxph_subgroups_df <- function(variables,
     tte = data[[variables$tte]],
     is_event = data[[variables$is_event]],
     arm = data[[variables$arm]],
-    strata_data = if (is.null(variables$strat)) NULL else data[variables$strat],
+    strata_data = if (is.null(variables$strata)) NULL else data[variables$strata],
     control = control
   )
   result_all$subgroup <- label_all
@@ -365,7 +374,7 @@ h_coxph_subgroups_df <- function(variables,
         tte = grp$df[[variables$tte]],
         is_event = grp$df[[variables$is_event]],
         arm = grp$df[[variables$arm]],
-        strata_data = if (is.null(variables$strat)) NULL else grp$df[variables$strat],
+        strata_data = if (is.null(variables$strata)) NULL else grp$df[variables$strata],
         control = control
       )
       result_labels <- grp$df_labels[rep(1, times = nrow(result)), ]